// services/api/types.ts
export interface EpisodeItem {
  guid: string
  title: string
  pubDate: string
  link: string
  author: string
  thumbnail: string
  description: string
  content: string
  enclosure: {
    link: string
    type: string
    length: number
    duration: number
    rating: { scheme: string; value: string }
  }
  categories: string[]
}